third_party/WebKit/LayoutTests/svg/dynamic-updates/SVGPathElement-valid-arguments.html

<!DOCTYPE html>
<html>
  <body>
    <script src="../../resources/js-test.js"></script>
    <script>
      var pathElement = document.createElementNS("http://www.w3.org/2000/svg","path");

      shouldThrow('pathElement.getPointAtLength()',
        '"TypeError: Failed to execute \'getPointAtLength\' on \'SVGPathElement\': 1 argument required, but only 0 present."');
      shouldThrow('pathElement.getPointAtLength(NaN)',
        '"TypeError: Failed to execute \'getPointAtLength\' on \'SVGPathElement\': The provided float value is non-finite."');
      shouldThrow('pathElement.getPointAtLength(Infinity)',
        '"TypeError: Failed to execute \'getPointAtLength\' on \'SVGPathElement\': The provided float value is non-finite."');

      shouldThrow('pathElement.getPathSegAtLength()',
      	'"TypeError: Failed to execute \'getPathSegAtLength\' on \'SVGPathElement\': 1 argument required, but only 0 present."');
      shouldThrow('pathElement.getPathSegAtLength(NaN)',
        '"TypeError: Failed to execute \'getPathSegAtLength\' on \'SVGPathElement\': The provided float value is non-finite."');
      shouldThrow('pathElement.getPathSegAtLength(Infinity)',
        '"TypeError: Failed to execute \'getPathSegAtLength\' on \'SVGPathElement\': The provided float value is non-finite."');
    </script>
	<p id="description">Test that correct exceptions are thrown from SVGPathElement methods.</p>
	<div id="console"></div>
  </body>
</html>
