third_party/WebKit/LayoutTests/svg/dynamic-updates/SVGPathElement-valid-arguments-expected.txt

PASS pathElement.getPointAtLength() threw exception TypeError: Failed to execute 'getPointAtLength' on 'SVGPathElement': 1 argument required, but only 0 present..
PASS pathElement.getPointAtLength(NaN) threw exception TypeError: Failed to execute 'getPointAtLength' on 'SVGPathElement': The provided float value is non-finite..
PASS pathElement.getPointAtLength(Infinity) threw exception TypeError: Failed to execute 'getPointAtLength' on 'SVGPathElement': The provided float value is non-finite..
PASS pathElement.getPathSegAtLength() threw exception TypeError: Failed to execute 'getPathSegAtLength' on 'SVGPathElement': 1 argument required, but only 0 present..
PASS pathElement.getPathSegAtLength(NaN) threw exception TypeError: Failed to execute 'getPathSegAtLength' on 'SVGPathElement': The provided float value is non-finite..
PASS pathElement.getPathSegAtLength(Infinity) threw exception TypeError: Failed to execute 'getPathSegAtLength' on 'SVGPathElement': The provided float value is non-finite..
PASS successfullyParsed is true

TEST COMPLETE
Test that correct exceptions are thrown from SVGPathElement methods.
